% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.species.R
\name{convert.species}
\alias{convert.species}
\title{Conversion of an SRL from one species to another.}
\usage{
convert.species(SRL.df, digest.df, SRL.format = "peakview")
}
\arguments{
\item{SRL.df}{A data frame of the spectral reference library from
\code{\link{import.srl}}.}

\item{digest.df}{A data frame of the digested protein sequence from the
desired species generated from  \code{\link{digest.fasta}}.}

\item{SRL.format}{The format of the SRL. Either "peakview" for PeakView or
OneOmics SRLs or "openswath" for OpenSWATH SRLs. Defaults to "peakview".}
}
\value{
A data frame of the new spectral reference library with the peptides
from the desired species.
}
\description{
Converts the species in the spectral reference library to the species in the
digested fasta data frame where full peptide identity occurs.
}
\note{
This process will remove the retention time calibration protein from a
PeakView/OneOmics SRL.
}
\examples{
data(rat_srl_example)
data(human_digest_example)
#convert.species(human_digest_example, rat_srl_example)


}
\seealso{


Other related: \code{\link{digest.fasta}},
  \code{\link{export.srl}}, \code{\link{import.fasta}},
  \code{\link{import.srl}}
}
\author{
Madeleine J Otway \email{motway@cmri.org.au}
}
