% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_species.R
\name{convertSpecies}
\alias{convertSpecies}
\title{Conversion of an SRL from one species to another.}
\usage{
convertSpecies(digest.df = NULL, SRL.df = NULL,
  SRL.format = "peakview")
}
\arguments{
\item{digest.df}{A data frame of the digested protein sequence from the
desired species generated from \code{\link{digestFasta}}.}

\item{SRL.df}{A data frame of the spectral reference library from
\code{\link{importSRL}}.}

\item{SRL.format}{The format of the SRL. Either "peakview" for PeakView or
OneOmics SRLs or "openswath" for OpenSWATH SRLs. Defaults to "peakview".}
}
\value{
A data frame of the new spectral reference library with the peptides
from the desired species.
}
\description{
Converts the species in the spectral reference library to the species in the
digested fasta data frame where full peptide identity occurs.
}
\note{
This process will remove the retention time calibration protein from a
PeakView/OneOmics SRL.
}
\examples{
data(rat_srl_example)
data(human_digest_example)
convertSpecies(human_digest_example, rat_srl_example)
convertSpecies(human_digest_example, rat_srl_example, SRL.format = "peakview")

## Workflow
fasta <- system.file("extdata",
                     "human_proteome_example.fasta",
                     package = "dialects")
human_proteome_example <- import.fasta(fasta)
human_digest_example <- digest.fasta(human_proteome_example)
srl_pv <- system.file("extdata",
                      "rat_srl_example.txt",
                      package = "dialects")
rat_srl_example <- import.srl(srl_pv, SRL.format = "peakview")
convertSpecies(human_digest_example, rat_srl_example)


}
\seealso{
For required functions before converting SRL, see:
\code{\link[dialects]{importFfasta}},
\code{\link{digestFasta}}, \code{\link{importSRL}}
}
\author{
Madeleine J Otway \email{motway@cmri.org.au}
}
