% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.fasta.R
\name{digest.fasta}
\alias{digest.fasta}
\title{Tryptic digestion of a protein sequence dataframe.}
\usage{
digest.fasta(fasta.df)
}
\arguments{
\item{fasta.df}{Data frame generated from \code{\link{import.fasta}}}
}
\value{
Returns a data frame of the peptide sequences of the digested
  proteins.
}
\description{
This function performs an in-silico trypsin digestion on a data frame from a
protein sequence fasta file.
}
\details{
Data frame must have the vector names: accession, name, sequence
}
\note{
This function removes all non-unique peptides. Removes all peptides
  less than 5 and greater than 52 amino acids.
}
\examples{
data(human_proteome_example)
digest.fasta(human_proteome_example)

## Workflow
fasta <- system.file("extdata",
                     "human_proteome_example.fasta",
                     package = "dialects")
human_proteome_example <- import.fasta(fasta)
digest.fasta(human_proteome_example)

}
\seealso{
For required functions before digesting fasta file, see:
\code{\link[dialects]{import.fasta}}
}
\author{
Madeleine J Otway \email{motway@cmri.org.au}
}
