% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_digest_example}
\alias{human_digest_example}
\title{Tryptic digestion of human proteome}
\format{A data frame containing 501 rows and 3 variables.
\describe{
  \item{accession}{(character) - UniProt accession code for a protein}
  \item{name}{(character) - Name a protein (as defined by UniProt)}
  \item{sequence}{(character) - Sequence of a trypticly digested peptide}
  }}
\source{
In silico digestion performed at the Children's Medical Research Intitute in 2017 by Madeleine J Otway.
}
\usage{
human_digest_example
}
\description{
This dataset contains an in silico tryptic digestion of 7 human proteins
(generated in \code{\link{digest.fasta}}). This tryptic digestion is for example
purposes only.
}
\details{
Each row corresponds to a single trypticly digested peptide.
}
\keyword{datasets}
