% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_proteome_example}
\alias{human_proteome_example}
\title{Tryptic digestion of human proteome}
\format{A data frame containing 7 rows and 3 variables.
\describe{
  \item{accession}{(character) - UniProt accession code for a protein}
  \item{name}{(character) - Name a protein (as defined by UniProt)}
  \item{sequence}{(character) - Sequence of a protein}
  }}
\source{
Proteome downloaded from \url{www.uniprot.org/} at the Children's Medical Research Intitute in 2017 by Madeleine J Otway.
}
\usage{
human_proteome_example
}
\description{
This dataset contains 7 human proteins (downloaded from \url{www.uniprot.org/}).
This dataset is for example purposes only.
}
\details{
Each row corresponds to a protein.
}
\keyword{datasets}
