% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.fasta.R
\name{import.fasta}
\alias{import.fasta}
\title{Import a fasta file.}
\usage{
import.fasta(fasta.file)
}
\arguments{
\item{fasta.file}{Directory and filename of fasta formatted protein sequence.}
}
\description{
Imports a fasta formatted protein sequence file (downloaded from
\url{www.uniprot.org}). This function is compatable with SwissProt and
TrEMBL entries.
}
\examples{
fasta <- system.file("extdata", "FASTA", package = "dialects")
import.fasta(fasta)


}
\seealso{


Other related: \code{\link{convert.species}},
  \code{\link{digest.fasta}}, \code{\link{export.srl}},
  \code{\link{import.srl}}
}
\author{
Madeleine J Otway \email{motway@cmri.org.au}
}
