% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.srl.R
\name{import.srl}
\alias{import.srl}
\title{Import an SRL.}
\usage{
import.srl(filepath, SRL.format = "peakview")
}
\arguments{
\item{filepath}{Directory and filename of the SRL.}

\item{SRL.format}{The format of the SRL. Either "peakview" for PeakView or
OneOmics SRLs or "openswath" for OpenSWATH SRLs. Defaults to "peakview".}
}
\description{
This function imports a spectral reference library (SRL) in either
PeakView/OneOmics or OpenSWATH formats.
}
\examples{
srl_pv <- system.file("extdata", "rat_srl_example.txt", package = "dialects")
import.srl(srl_pv)
import.srl(srl_pv, SRL.format = "peakview")

#srl-os <- system.file("extdata", "SRL-OS", package = "dialects")
#import.srl(srl-os, SRL.format = "openswath")


}
\seealso{


Other related: \code{\link{convert.species}},
  \code{\link{digest.fasta}}, \code{\link{export.srl}},
  \code{\link{import.fasta}}
}
\author{
Madeleine J Otway \email{motway@cmri.org.au}
}
