% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fasta.R
\name{importFasta}
\alias{importFasta}
\title{Import a fasta file.}
\usage{
importFasta(fasta.file = NULL)
}
\arguments{
\item{fasta.file}{Full path to filename of fasta formatted protein sequence.}
}
\value{
Fasta sequences
}
\description{
Imports a fasta formatted protein sequence file (downloaded from
\url{www.uniprot.org}). This function is compatable with SwissProt and TrEMBL
entries.
}
\examples{
fasta <- system.file("extdata",
                     "human_proteome_example.fasta",
                     package = "dialects")
importFasta(fasta)


}
\seealso{
To perform an \emph{in silico} tryptic digestion of the imported
fasta file, see: \code{\link[dialects]{digestFasta}}
}
\author{
Madeleine J Otway \email{motway@cmri.org.au}
}
