% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_srl.R
\name{importSRL}
\alias{importSRL}
\title{Import an SRL.}
\usage{
importSRL(filepath, SRL.format = "peakview")
}
\arguments{
\item{filepath}{Directory and filename of the SRL.}

\item{SRL.format}{The format of the SRL. Either "peakview" for PeakView or
OneOmics SRLs or "openswath" for OpenSWATH SRLs. Defaults to "peakview".}
}
\value{
SRL
}
\description{
This function imports a spectral reference library (SRL) in either
PeakView/OneOmics or OpenSWATH formats.
}
\note{
PeakView and OneOmics SRLs must be .txt format. OpenSWATH must be either .tsv or .csv format.
}
\examples{
srl_pv <- system.file("extdata",
                      "rat_srl_example.txt",
                      package = "dialects")
importSRL(srl_pv)
importSRL(srl_pv, SRL.format = "peakview")

#srl-os <- system.file("extdata", "SRL-OS", package = "dialects")
#importSRL(srl-os, SRL.format = "openswath")

}
\author{
Madeleine J Otway \email{motway@cmri.org.au}
}
