% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.species.R
\name{convert.species}
\alias{convert.species}
\title{Conversion of an SRL from one species to another.}
\usage{
convert.species(SRL.df, digest.df)
}
\arguments{
\item{SRL.df}{A data frame of the spectral reference library from
\code{\link{import.srl}}}

\item{digest.df}{A data frame of the digested protein sequence from the
desired species generated from  \code{\link{digest.fasta}}}
}
\value{
A data frame of the new spectral reference library with the peptides
from the desired species.
}
\description{
Converts the species in the spectral reference library to the species in the
digested fasta data frame where full peptide identity occurs.
}
\examples{
digest <- system.file("extdata", "DIGEST", package = "dialects")
srl <- ("extdata", "SRL", package = "dialects")
convert.species(digest, srl)


}
\seealso{


Other related: \code{\link{digest.fasta}},
  \code{\link{export.srl}}, \code{\link{import.fasta}},
  \code{\link{import.srl}}
}
\author{
Madeleine J Otway \email{motway@cmri.org.au}
}
