% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.fasta.R
\name{digest.fasta}
\alias{digest.fasta}
\title{Tryptic digestion of a protein sequence dataframe.}
\usage{
digest.fasta(fasta.df)
}
\arguments{
\item{fasta.df}{Data frame generated from \code{\link{import.fasta}}}
}
\value{
Returns a data frame of the peptide sequences of the digested
  proteins.
}
\description{
This function performs an in-silico trypsin digestion on a data frame from a
protein sequence fasta file.
}
\details{
Data frame must have the vector names: accession, name, sequence
}
\note{
This function removes all non-unique peptides. Removes all peptides
  less than 5 and greater than 52 amino acids.
}
\examples{
load("~/dialects/data/human_proteome_example.rda")
digest.fasta(human_proteome_example)


}
\seealso{


Other related: \code{\link{convert.species}},
  \code{\link{export.srl}}, \code{\link{import.fasta}},
  \code{\link{import.srl}}
}
\author{
Madeleine J Otway \email{motway@cmri.org.au}
}
