% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.srl.R
\name{export.srl}
\alias{export.srl}
\title{Export an SRL.}
\usage{
export.srl(SRL.df, SRL.filepath)
}
\arguments{
\item{SRL.df}{Data frame of an SRL}

\item{SRL.filepath}{The directory and filename of the SRL to be exported}
}
\description{
Exports a spectral reference library (SRL) in either PeakView/OneOmics or
OpenSWATH formats.
}
\examples{
srl <- ("extdata", "SRL", package = "dialects")
export.srl(srl, "new-srl.txt")


}
\seealso{


Other related: \code{\link{convert.species}},
  \code{\link{digest.fasta}}, \code{\link{import.fasta}},
  \code{\link{import.srl}}
}
\author{
Madeleine J Otway \email{motway@cmri.org.au}
}
