% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_srl.R
\name{exportSRL}
\alias{exportSRL}
\title{Export an SRL.}
\usage{
exportSRL(SRL.df, SRL.filepath)
}
\arguments{
\item{SRL.df}{Data frame of an SRL}

\item{SRL.filepath}{The directory and filename of the SRL to be exported}
}
\description{
Exports a spectral reference library (SRL) in either PeakView/OneOmics or
OpenSWATH formats.
}
\examples{
data(rat_srl_example)
exportSRL(rat_srl_example, "new-srl.txt")

## Entire Workflow
fasta <- system.file("extdata",
                     "human_proteome_example.fasta",
                     package = "dialects")
human_proteome_example <- import.fasta(fasta)
human_digest_example <- digest.fasta(human_proteome_example)
srl_pv <- system.file("extdata",
                      "rat_srl_example.txt",
                      package = "dialects")
rat_srl_example <- import.srl(srl_pv, SRL.format = "peakview")
human_from_rat <- convert.species(human_digest_example, rat_srl_example)
exportSRL(human_from_rat, "human_from_rat_srl.txt")


}
\seealso{
For functions require to convert SRL before exporting SRL, see:
\code{\link[dialects]{import.fasta}}, \code{\link{digest.fasta}},
\code{\link{import.srl}}, \code{\link{convert.species}}
}
\author{
Madeleine J Otway \email{motway@cmri.org.au}
}
