% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.fasta.R
\name{import.fasta}
\alias{import.fasta}
\title{Import a fasta file.}
\usage{
import.fasta(fasta.file)
}
\arguments{
\item{fasta.file}{Directory and filename of fasta formatted protein sequence.}
}
\description{
Imports a fasta formatted protein sequence file (downloaded from
\url{www.uniprot.org}). This function is compatable with SwissProt and TrEMBL
entries.
}
\examples{
fasta <- system.file("extdata",
                     "human_proteome_example.fasta",
                     package = "dialects")
import.fasta(fasta)


}
\seealso{
To perform an \emph{in silico} tryptic digestion of the imported
fasta file, see: \code{\link[dialects]{digest.fasta}}
}
\author{
Madeleine J Otway \email{motway@cmri.org.au}
}
